requirements.txt
Django~=2.2
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
PyJWT==1.7.1
pytz==2019.1